requirements.txt
amqp==2.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.0.0
certifi==2018.1.18
chardet==3.0.4
Django==1.9.5
djangorestframework==3.3.3
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
Pillow==3.2.0
psycopg2==2.7.4
py==1.5.2
pytest==2.9.1
python-gcm==0.4
pytz==2018.3
redis==2.10.5
requests==2.18.4
urllib3==1.22
vine==1.1.4
